add recent changes
Backfit #2581 #2583 for osm in www_base

Could(should?) of had the files live in the osm_vector_maps role or in the map
git repo adding the needed fetch to install.yml. A quick "reinstall" of the
role would of been possible. Could added stanzas to copy the new files into
place or just repeat the role. Less typing to repeat the role but not the
fastest way.

Not really needed as there is no check in place for 'installed' within www_base

nodejs - www_base correct <= to <

handle via iiab-upgrade script

update notes

more updates Nov3