Fix naturalearth symlink
Previously, we erroneously pointed the openstretmap pmtiles symlink to
the naturalearth pmtiles file. I think this happened to be okay because
the frontend requested the naturalearth file directly anyway, and it was
there. However now that we have the date, we need to point the
non-date path to the actual file which has the date.