--show-console-readout=false again for aria2
We don't want a bunch of those status lines that goes to the console
that gets overwritten, but since we're piping it to a log file it just
gets accumulated. It's way more common than once a minute.

But summaries still work. Turns out summaries don't count as "console readout".